Started testing_utilities module: implemented call_should_fail function.
Started a module containing testing utilities, with a single function
for now.

The call_should_fail function is for making sure a given function raises
an exception with given arguments: make sure that funtion(*args) raises
Error; work silently if it does, raise NewError(message) if it doesn't.

When the module is run directly, it tests the functionality.