Added keybased_defaultdict class.
keybased_defaultdict is a defaultdict equivalent that passes the key as
an argument to the default-value factory, instead of no argumnets.

Normal defaultdict(int)[9] is 0, because no argument is passed to the
int function, and int() is 0.  On the other hand
keybased_defaultdict(int)[9] would be 9, keybased_defaultdict(bool)[9]
would be True, etc.

From
http://stackoverflow.com/questions/2912231/is-there-a-clever-way-to-pass-the-key-to-defaultdicts-default-factory