Added FrozenClass: mix-in to disallow adding new attributes after creation.
Class that allows prevention of adding new attributes after creation.
Use by inheriting from FrozenClass when creating a new class you want to
"freeze", and then running self._freeze() at the end of __init__, like this:
    class Test_freezing(FrozenClass):
        def __init__(self, x, y):
            self.x = x
            self.y = y
            self._freeze() # no new attributes after this point.
    a = Test_freezing(2,3)
    a.x = 10    # can modify existing attributes after creation
    a.z = 10    # fails - cannot add NEW atributes after creation

I got the code from
http://stackoverflow.com/questions/3603502/prevent-creating-new-attributes-outside-init
(this is also stated in the comments).

MAYBE-TODO - could/should this be a decorator instead?

Added unittest code for it - it passes.