Added an option to write output to file instead of to stdout.
In read_length_distribution and print_lengths, instead of always
printing output to stdout using print, I used equivalent OUTPUT.write
statements, with OUTPUT as a function argument, sys.stdout by default.
That way I can make the output go to a file instead of to stdout, if
running the function from python, not straight from the command-line.

Tested both output versions (stdout and to file).

TODO it'd be even more convenient to be able to get the output either to
stdout/file or as a returned list of lines (or both, I suppose), but
that might be a bit much...  I'll see.