Added simple count_list_values function (returns value:count dict).
Added a four-line count_list_values function - Given a list, return a
value:number_of_times_value_occurred_in_list dictionary.  It's
surprising that's not in itertools or such already...  Itertools.groupby
is the closest, but it returns an odd iterator of value:iterator items,
and it requires a sort first, so it's both slower and more awkward to
use.  Collections.Counter is basically this with more features, but it
was added for python 2.7, and I'm on 2.6, so I have to use my own
version for now.

Wrote a unit-test for the function; it passes.