Added force_fasta_output option; outfile name format changes.
Added an option to make the output fasta even if the input was fastq
(easy).  Made sure to also change the outfile extension to 'fa' when
doing that.

Added "bp" to the output filenames, and changed the the
optional min/max ones from "lower/higher" to "underXbp/overXbp".

Replaced main function docstring with reference to module docstring and
option help messages to avoid duplication of information.