Added option to discard zero-length seqs; made printing to stdout work.
Added an option to discard zero-length sequences instead of printing
them to a file, since that file might be useless, and whether a file
with no sequences and just headers is a valid fasta-format file is
dubious (less dubious for fastq).  Tested, works.

Also fixed a lot of bugs that were making printing stuff to stdout not
work - works now.