Function to compare two files with optional regex lines (for functional tests).
Started new file testing_utilities.py, for utilities related to
unit testing, functional testing etc.

Wrote one function - compare_files_with_regex.
The reason for this function is to be able to compare an output file and
a reference file but ignore things that are expected to change (like
the date in the header) - the reference file can contain lines starting
with <REGEX> and then giving a regular expression (.* is fine), in
which case the output file lines are compared against the regex instead,
which allows wildcards for dates/etc.  (See my stackoverflow question
http://stackoverflow.com/questions/9726214/testing-
full-program-by-comparing-output-file-to-reference-file-whats-it-calle)
Function docstring:
    Given two string iterators, return True if they're identical except
    for special treatment of <REGEX> and <IGNORE>.
    The inputs can be lists of strings, generators, sets, open files,
    etc.  NO INPUT CHECKING IS DONE.
    The program goes over both iterators in parallel (always
    looking at element1 and element2), as follows:
     - if either element starts with <IGNORE>, ignore it and advance its
       iterator to the next element.
     - if both elements start with <REGEX>, raise an exception.
     - if one element starts with <REGEX>, parse the rest of it as a
       regular expression: (removing whitespace from start/end, and
       adding ^ and $ to start and end) if the other element matches the
       regex, advance both iterators, otherwise return False
     - if neither element starts with either special string, compare
       them directly: if the elements are identical, advance both
       iterators, otherwise return False If the end of both iterators
       has been reached without finding a mismatch, return True.
     If ignore_empty_lines is True, all lines containing nothing or just
     whitespace are ignored (treated like <IGNORE>).  If
     ignore_whitespace is True, change all sequences of multiple
     spaces/tabs to a single space, and remove all space at the
     beginning and end, before doing comparisons.  If ignore_case is
     True, change all elements to lowercase before doing comparisons.

(Also wrote two help functions for that, _advance_iter_keep_state and
clean_line)

Wrote a lot of unit-tests for the function, including ones that actually
take input files (the new test_inputs/textcmp_file*.txt files) instead
of just lists.  They all pass.

Missing things to possibly add later, but not worth it now:
 - input checking for the main function (and unit-tests for it).
 - unit-tests for the help functions.