Made compare_files_with_regex return useful info if comparison fails.
Originally compare_files_with_regex just returned True if the files
matched and False if they didn't, but that's not very practical when you
have to find the difference between large files by eye to find the
error, because just running them through diff will find all of the
special lines etc, which can't be the problem.  So instead, not True is
returned if the files match, and if they don't match, the return is a
tuple of the mismatched lines (or, if one file ended and the other one
had lines left, a tuple composed of a string explaining what happened
and the next line of the longer file).

Modified the docstring to state that, and also rewrote it a bit in
general.
Modified the unit-test lines for failed matches to reflect the new
behavior:
 - in some cases I tested the specific return value for failed matches
 - in most cases I didn't test the specific value, just !=True.