Bug fix for compare_files_with_regex message when one file is longer.
I added a unit-test checking the specific return value in the
non-matching case when one input has more lines than the other (the
return should be a tuple of an info message and the next line in the
longer input), and originally it failed.
The reason was how the while-loop was set up - after one of the
iterators stopped, more stuff was done (including a comparison of the
fake empty line to the other line) instead of the loop being exited
immediately, which gave the wrong error message type.

Fixed that by adding an explicit break condition right after the two
_advance_iter_keep_state calls (and changed the loop to a "while True"
loop, since nothing more complicated is necessary now).

All unit-tests pass.