Modified parse_fasta.py to not skip empty sequences; other comments/TODOs.
Original version of parse_fasta.py skipped empty sequences when reading,
which turned out to be a problem, because sometimes there are empty
sequences in files (for example in cutadapt output if the "insert" was
actually just two adapters ligated together and we stripped both of
them).  So I modified it to not skip empty lines and to not require a
non-empty sequence for printing.  Also modified print_seq to have a more
sensible format.

I tested the new version by hand on the test_inputs/test.fa file.