Cleaning up imports; moving optional ones away from file top.
Moved imports that are only used in "if __name__ == '__main__':"
section to that section, to allow the module to be imported without
needing the dependency.  It's not perfect python style, but it'll be
more convenient, I think.

Removed "from read_input import read_input" from general_utilities.py -
it was annoying when one of the programs I have to someone else imported
general_utilities.py and then it turned out it needed read_input to
despite not using it.

Removed a few unused and duplicate imports.