Minor: changed order of files for error message consistency.
Changed "compare_files_with_regex(OUTFILE, REFFILE)" to
"compare_files_with_regex(REFFILE, OUTFILE)" to make the order of the
mismatched lines returned as an error message match the order in which
the two files are listed in the error message below (reference files
have more meaningful names, so it makes sense for the reference file to
be first).

compare_files_with_regex is fully symmetrical, so this won't change
anything.  Unit-tests pass.  Added MAYBE-TODO comment to add some
unit-tests to make sure it really is symmetrical, though.