Added value_and_percentages function: 1,[2,3] -> "1 (50%, 33%)".
Return a string containing val and val as percentage of each total:
1,[2,3] -> "1 (50%, 33%)".

If fractions_not_percentages=True, 1,[2,3] -> "1 (0.5, 0.33)" instead.
percentage_format_str is the string used to format each
percentage/fraction (in .X, X is the precision in digits).
If exception_for_100 is True, 100 is formatted as '100' rather than
'1e+02' even if precision is 2.
If exception_for_100 is 'default', it's True for percentages with '%.2g'
format but False otherwise and for fractions.

Added unit-tests, they all pass.