Tricky format_base_distance bug due to old-style division!
There was a tricky bug in format_base_distance, where the unit-tests
passed, but if it was run from inside something with "from __future__
import division", the results in the no-approximation variant were
floats (so 1000 was 1.0kb instead of 1kb, etc).
Added "from __future__ import division" to this file (all files should
have it, this one was just old!), which caused some unit-tests to fail,
so I fixed the format_base_distance function, then had to fix some other
unit-tests due to the same problem, and added more direct unit-tests
just in case.  All pass.