Inverted GC_content other-bases argument for clarity; changed error messages.
In GC_content, changed the error_other_bases argument to
ignore_other_bases (the opposite) for clarity; changed the code
and unit-tests to match.  No functionality change besides inverting the
argument value.

Also changed the error messages to NOT print the sequence, because
sometimes it's really long (whole genome etc) and we really don't want
to do that.