Modified savefig to always use extensions if given.
Old behavior: if filename has a '.', use that; otherwise if extensions
given (default png), use those, otherwise error.
New behavior: if extensions given (default None), use them; otherwise if
filename has a '.', use that, otherwise use png.
New behavior makes more sense really, and implemented more cleanly; it
just required the 'png' default to be given in the code rather than as
an argument default, so we can tell if the user provided 'png' or if it
was a default.
Did this because I was trying to use savefig on something like
figure_linewidth-0.5 and it was interpreting the .5 as an extension.